leetcode/src/main/java/com/algorithmic/problems/P110_BalancedBinaryTree.java
package com.algorithmic.problems;

import com.algorithmic.model.TreeNode;

public class P110_BalancedBinaryTree {

    public boolean isBalanced(TreeNode root) {
        if (root == null) {
            return true;
        }
        int absDiff = Math.abs(height(root.left) - height(root.right));
        if (absDiff > 1) {
            return false;
        } else {
            return isBalanced(root.left) && isBalanced(root.right);
        }
    }

    private int height(TreeNode root) {
        if (root == null) {
            return 0;
        }
        return Math.max(height(root.left), height(root.right)) + 1;
    }
}
